refactor: divide styles into several files
