+++ src/App.scss
@@ -58,114 +58,8 @@ input {
   width: 100%;
 }
 
-.page {
-  display: flex;
-}
-
-.main-page {
-  width: 100%;
-}
-
-.top-section {
-  padding: 30px;
-  background-color: #1a1a1a;
-}
-
-.error-button {
-  position: absolute;
-  top: 20px;
-  left: 20px;
-}
-
 .hiding {
   display: none !important;
-}
-
-.search-block {
-  width: 250px;
-  margin: 0 auto;
-  color: white;
-}
-
-.search {
-  display: flex;
-  justify-content: space-between;
-  margin-top: 10px;
-  background-color: rgb(70, 70, 70);
-  border-radius: 10px;
-  border: solid 2px #fddd32;
-}
-
-.search-input {
-  color: white;
-  border: none;
-  background-color: rgb(70, 70, 70);
-  border-radius: 10px;
-}
-
-.loupe {
-  width: 20px;
-  height: 30px;
-  margin: 5px 8px 0;
-  background-image: url('../src/assets/drawing-2.svg');
-  background-size: contain;
-  background-repeat: no-repeat;
-  &:hover {
-    cursor: pointer;
-  }
-}
-
-.bottom-section {
-  background-color: #bf7232;
-  height: 100vh;
-  padding-top: 20px;
-}
-
-.loading {
-  display: flex;
-  flex-wrap: wrap;
-  align-content: center;
-  width: 150px;
-  margin: 0 auto 40px;
-  div {
-    width: 15px;
-    margin-right: 15px;
-    height: 15px;
-    animation: loading linear 1.6s infinite;
-  }
-  @for $i from 1 through 5 {
-    div:nth-child(#{$i}) {
-      animation-delay: 0 + 0.2s * ($i - 1);
-    }
-  }
-}
-
-.bubble {
-  background-image: url('../src/assets/PIWO.svg');
-  background-size: contain;
-}
-
-.not-found {
-  text-align: center;
-  color: black;
-}
-
-.result-container {
-  display: grid;
-  grid-template-columns: repeat(auto-fit, 250px);
-  gap: 10px;
-  justify-content: space-evenly;
-  width: 95%;
-  margin: 0 auto;
-}
-
-.result-item {
-  padding: 0 10px 10px;
-  border: 2px solid rgb(120, 120, 120);
-  border-radius: 5px;
-  font-size: 12px;
-  background-color: rgb(239, 236, 174);
-  position: relative;
 }
 
 .error-block {
@@ -204,51 +98,4 @@ input {
   border-radius: 50%;
   background-color: white;
   animation: looking 4s infinite;
-}
-
-.pagination {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  font-size: 55px;
-  div:nth-child(2) {
-    font-size: 20px;
-    text-align: center;
-    font-weight: 700;
-    border-radius: 100%;
-    border: 2px solid #213547;
-    padding: 1px 10px;
-  }
-  &:hover {
-    cursor: pointer;
-  }
-}
-
-.card-checkbox {
-  position: absolute;
-  top: 10px;
-  right: 10px;
-  visibility: hidden;
-}
-
-.card-checkbox::before {
-  content: '✓';
-  flex-wrap: wrap;
-  display: flex;
-  align-content: center;
-  justify-content: center;
-  visibility: visible;
-  height: 11px;
-  width: 11px;
-  position: absolute;
-  border: 1px solid #213547;
-  border-radius: 2px;
-  background-color: rgb(243 242 213);
-  color: rgb(243 242 213);
-  font-size: 12px;
-}
-
-.card-checkbox:checked::before {
-  background-color: #213547;
-  color: white;
 }
