+++ src/components/404/404.tsx
@@ -3,7 +3,7 @@ export default function NotFound() {
     <>
       <div className="not-found">
         <h1>404</h1>
-        <h3>Page noy found</h3>
+        <h3>Page not found</h3>
       </div>
     </>
   );
