+++ src/components/card/card.scss
@@ -1,0 +1,37 @@
+.result-item {
+  padding: 0 10px 10px;
+  border: 2px solid rgb(120, 120, 120);
+  border-radius: 5px;
+  font-size: 12px;
+  background-color: rgb(239, 236, 174);
+  position: relative;
+}
+
+.card-checkbox {
+  position: absolute;
+  top: 10px;
+  right: 10px;
+  visibility: hidden;
+}
+
+.card-checkbox::before {
+  content: '✓';
+  flex-wrap: wrap;
+  display: flex;
+  align-content: center;
+  justify-content: center;
+  visibility: visible;
+  height: 11px;
+  width: 11px;
+  position: absolute;
+  border: 1px solid #213547;
+  border-radius: 2px;
+  background-color: rgb(243 242 213);
+  color: rgb(243 242 213);
+  font-size: 12px;
+}
+
+.card-checkbox:checked::before {
+  background-color: #213547;
+  color: white;
+}
