+++ src/components/flyout/flyout.scss
@@ -7,3 +7,13 @@
   color: white;
   padding: 15px;
 }
+
+.flyout__title {
+  text-align: center;
+  margin-top: 0;
+}
+
+.flyout__buttons {
+  display: flex;
+  justify-content: space-between;
+}
