+++ src/components/loader/loader.tsx
@@ -1,3 +1,5 @@
+import './loader.scss';
+
 export function Loader() {
   return (
     <div className="loading">
