+++ src/components/page/page.scss
@@ -1,0 +1,18 @@
+.page {
+  display: flex;
+}
+
+.main-page {
+  width: 100%;
+}
+
+.top-section {
+  padding: 30px;
+  background-color: #1a1a1a;
+}
+
+.bottom-section {
+  background-color: #bf7232;
+  height: 100vh;
+  padding-top: 20px;
+}
