+++ src/components/pagination/pagination.tsx
@@ -1,6 +1,7 @@
 import { useSearchParams } from 'react-router-dom';
 import { useSelector } from 'react-redux';
-import { RootState } from '../store/store';
+import { RootState } from '../../store/store';
+import './pagination.scss';
 
 export default function Pagination() {
   const [searchParams, setSearchParams] = useSearchParams();
