+++ src/components/search/search.scss
@@ -1,0 +1,33 @@
+.search-block {
+  width: 250px;
+  margin: 0 auto;
+  color: white;
+}
+
+.search {
+  display: flex;
+  justify-content: space-between;
+  margin-top: 10px;
+  background-color: rgb(70, 70, 70);
+  border-radius: 10px;
+  border: solid 2px #fddd32;
+}
+
+.search-input {
+  color: white;
+  border: none;
+  background-color: rgb(70, 70, 70);
+  border-radius: 10px;
+}
+
+.loupe {
+  width: 20px;
+  height: 30px;
+  margin: 5px 8px 0;
+  background-image: url('../src/assets/drawing-2.svg');
+  background-size: contain;
+  background-repeat: no-repeat;
+  &:hover {
+    cursor: pointer;
+  }
+}
